# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server API",
    "version": "1.0.0",
    "description": "Machine-readable API contract for the Express backend in server."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyAuthScheme": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative format: `Authorization: ApiKey <apiKey>`"
      },
      "inboundBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for worker ingress"
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {},
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "EmailSender": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "type": "string",
                "nullable": true
              }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "Email": {
        "type": "object",
        "description": "Full email including text/html body, returned by the detail route.",
        "required": [
          "id",
          "to",
          "localPart",
          "domain",
          "envelopeFrom",
          "from",
          "subject",
          "workerName",
          "sourceDomain",
          "messageId",
          "receivedAt",
          "createdAt",
          "hasRawMime",
          "rawMimeSize",
          "text",
          "html",
          "preview",
          "hasText",
          "hasHtml"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "to": {
            "type": "string"
          },
          "localPart": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "envelopeFrom": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "$ref": "#/components/schemas/EmailSender"
          },
          "subject": {
            "type": "string"
          },
          "workerName": {
            "type": "string"
          },
          "sourceDomain": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "hasRawMime": {
            "type": "boolean"
          },
          "rawMimeSize": {
            "type": "integer"
          },
          "text": {
            "type": "string"
          },
          "html": {
            "type": "string"
          },
          "preview": {
            "type": "string",
            "description": "First 400 characters of the text body, used to render list rows"
          },
          "hasText": {
            "type": "boolean"
          },
          "hasHtml": {
            "type": "boolean"
          },
          "rawMime": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SimpleSuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "BlockedSenderMatch": {
        "type": "object",
        "description": "Present only when the inbound email was rejected by a blocked sender rule",
        "properties": {
          "id": {
            "type": "integer"
          },
          "patternType": {
            "type": "string",
            "enum": [
              "email",
              "domain"
            ]
          },
          "pattern": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "global",
              "domain"
            ]
          },
          "reason": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          }
        }
      },
      "EmailSummary": {
        "type": "object",
        "description": "Trimmed version used for lists; full text/html omitted so the payload does not grow with email count. Call GET /v1/emails/{id} to fetch the full content.",
        "required": [
          "id",
          "to",
          "localPart",
          "domain",
          "envelopeFrom",
          "from",
          "subject",
          "workerName",
          "sourceDomain",
          "messageId",
          "receivedAt",
          "createdAt",
          "hasRawMime",
          "rawMimeSize",
          "preview",
          "hasText",
          "hasHtml"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "to": {
            "type": "string"
          },
          "localPart": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "envelopeFrom": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "$ref": "#/components/schemas/EmailSender"
          },
          "subject": {
            "type": "string"
          },
          "workerName": {
            "type": "string"
          },
          "sourceDomain": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "hasRawMime": {
            "type": "boolean"
          },
          "rawMimeSize": {
            "type": "integer"
          },
          "preview": {
            "type": "string",
            "description": "First 400 characters of the text body, used to render list rows"
          },
          "hasText": {
            "type": "boolean"
          },
          "hasHtml": {
            "type": "boolean"
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/v1/emails": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "List emails",
        "description": "Global admins can list all emails with scope=system. Any caller can force scope=registered to only see emails whose mailbox is both permitted and registered in email_registers for that user. domain/address/search filters narrow the result set. search supports multiple terms and checks subject, body, stored metadata and raw_mime when available.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string",
              "enum": [
                "registered",
                "system"
              ]
            }
          },
          {
            "in": "query",
            "name": "domain",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "address",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "emails": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmailSummary"
                      }
                    },
                    "hasMore": {
                      "type": "boolean"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad query or missing scope filters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/emails/bulk-delete": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete emails in bulk",
        "description": "Deletes up to 200 emails in one request. Normal users must have write access to each selected email. The response includes deletedIds as well as missingIds and deniedIds so clients can reconcile stale selections.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "emailIds"
                ],
                "properties": {
                  "emailIds": {
                    "type": "array",
                    "maxItems": 200,
                    "items": {
                      "oneOf": [
                        {
                          "type": "integer",
                          "minimum": 1
                        },
                        {
                          "type": "string",
                          "minLength": 1
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk delete result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "deleted": {
                      "type": "integer"
                    },
                    "requestedCount": {
                      "type": "integer"
                    },
                    "deletedIds": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    },
                    "missingIds": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    },
                    "deniedIds": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/emails/{id}": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get email by ID",
        "description": "Normal users can only read emails whose mailbox is registered for them and still covered by permission. Global admins can read any email.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "includeRawMime",
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "$ref": "#/components/schemas/Email"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete email by ID",
        "description": "Normal users must have domain permission for the mailbox domain and the mailbox must be registered for them.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleSuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/inbound/email": {
      "post": {
        "tags": [
          "Inbound"
        ],
        "summary": "Ingest raw MIME email from worker",
        "security": [
          {
            "inboundBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "message/rfc822": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-Email-Envelope-To",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Envelope-From",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Worker-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Received-At",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Domain",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Email-Message-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Email accepted; blocked senders are acknowledged but discarded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "integer",
                      "nullable": true,
                      "description": "null when the email was blocked and therefore not stored"
                    },
                    "envelopeTo": {
                      "type": "string"
                    },
                    "domain": {
                      "type": "string"
                    },
                    "receivedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    },
                    "blocked": {
                      "type": "boolean",
                      "description": "true when a blocked sender rule discarded the email"
                    },
                    "blockedBy": {
                      "$ref": "#/components/schemas/BlockedSenderMatch"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}